Moved Control enum to the right namespace
