--- LemonUI/Elements/Control.cs
@@ -1,365 +1,0 @@
-﻿#if ALTV
-namespace LemonUI.Elements;
-
-public enum Control
-{
-    NextCamera = 0,
-    LookLr = 1,
-    LookUd = 2,
-    LookUpOnly = 3,
-    LookDownOnly = 4,
-    LookLeftOnly = 5,
-    LookRightOnly = 6,
-    CinematicSlowmo = 7,
-    ScriptedFlyUd = 8,
-    ScriptedFlyLr = 9,
-    ScriptedFlyZup = 10,
-    ScriptedFlyZdown = 11,
-    WeaponWheelUd = 12,
-    WeaponWheelLr = 13,
-    WeaponWheelNext = 14,
-    WeaponWheelPrev = 15,
-    SelectNextWeapon = 16,
-    SelectPrevWeapon = 17,
-    SkipCutscene = 18,
-    CharacterWheel = 19,
-    MultiplayerInfo = 20,
-    Sprint = 21,
-    Jump = 22,
-    Enter = 23,
-    Attack = 24,
-    Aim = 25,
-    LookBehind = 26,
-    Phone = 27,
-    SpecialAbility = 28,
-    SpecialAbilitySecondary = 29,
-    MoveLr = 30,
-    MoveUd = 31,
-    MoveUpOnly = 32,
-    MoveDownOnly = 33,
-    MoveLeftOnly = 34,
-    MoveRightOnly = 35,
-    Duck = 36,
-    SelectWeapon = 37,
-    Pickup = 38,
-    SniperZoom = 39,
-    SniperZoomInOnly = 40,
-    SniperZoomOutOnly = 41,
-    SniperZoomInSecondary = 42,
-    SniperZoomOutSecondary = 43,
-    Cover = 44,
-    Reload = 45,
-    Talk = 46,
-    Detonate = 47,
-    HudSpecial = 48,
-    Arrest = 49,
-    AccurateAim = 50,
-    Context = 51,
-    ContextSecondary = 52,
-    WeaponSpecial = 53,
-    WeaponSpecialTwo = 54,
-    Dive = 55,
-    DropWeapon = 56,
-    DropAmmo = 57,
-    ThrowGrenade = 58,
-    VehMoveLr = 59,
-    VehMoveUd = 60,
-    VehMoveUpOnly = 61,
-    VehMoveDownOnly = 62,
-    VehMoveLeftOnly = 63,
-    VehMoveRightOnly = 64,
-    VehSpecial = 65,
-    VehGunLr = 66,
-    VehGunUd = 67,
-    VehAim = 68,
-    VehAttack = 69,
-    VehAttack2 = 70,
-    VehAccelerate = 71,
-    VehBrake = 72,
-    VehDuck = 73,
-    VehHeadlight = 74,
-    VehExit = 75,
-    VehHandbrake = 76,
-    VehHotwireLeft = 77,
-    VehHotwireRight = 78,
-    VehLookBehind = 79,
-    VehCinCam = 80,
-    VehNextRadio = 81,
-    VehPrevRadio = 82,
-    VehNextRadioTrack = 83,
-    VehPrevRadioTrack = 84,
-    VehRadioWheel = 85,
-    VehHorn = 86,
-    VehFlyThrottleUp = 87,
-    VehFlyThrottleDown = 88,
-    VehFlyYawLeft = 89,
-    VehFlyYawRight = 90,
-    VehPassengerAimRight = 91,
-    VehPassengerAttack = 92,
-    VehSpecialAbilityFranklin = 93,
-    VehStuntUD = 94,
-    VehCinematicUD = 95,
-    VehCinematicUpOnly = 96,
-    VehCinematicDownOnly = 97,
-    VehCinematicLR = 98,
-    VehSelectNextWeapon = 99,
-    VehSelectPrevWeapon = 100,
-    VehRoof = 101,
-    VehJump = 102,
-    VehGrapplingHook = 103,
-    VehShuffle = 104,
-    VehDropProjectile = 105,
-    VehMouseControlOverride = 106,
-    VehFlyRollLR = 107,
-    VehFlyRollLeftOnly = 108,
-    VehFlyRollRightOnly = 109,
-    VehFlyPitchUD = 110,
-    VehFlyPitchUpOnly = 111,
-    VehFlyPitchDownOnly = 112,
-    VehFlyUndercarriage = 113,
-    VehFlyAttack = 114,
-    VehFlySelectNextWeapon = 115,
-    VehFlySelectPrevWeapon = 116,
-    VehFlySelectTargetLeft = 117,
-    VehFlySelectTargetRight = 118,
-    VehFlyVerticalFlightMode = 119,
-    VehFlyDuck = 120,
-    VehFlyAttackCamera = 121,
-    VehFlyMouseControlOverride = 122,
-    VehSubTurnLR = 123,
-    VehSubTurnLeftOnly = 124,
-    VehSubTurnRightOnly = 125,
-    VehSubPitchUD = 126,
-    VehSubPitchUpOnly = 127,
-    VehSubPitchDownOnly = 128,
-    VehSubThrottleUp = 129,
-    VehSubThrottleDown = 130,
-    VehSubAscend = 131,
-    VehSubDescend = 132,
-    VehSubTurnHardLeft = 133,
-    VehSubTurnHardRight = 134,
-    VehSubMouseControlOverride = 135,
-    VehPushbikePedal = 136,
-    VehPushbikeSprint = 137,
-    VehPushbikeFrontBrake = 138,
-    VehPushbikeRearBrake = 139,
-    MeleeAttackLight = 140,
-    MeleeAttackHeavy = 141,
-    MeleeAttackAlternate = 142,
-    MeleeBlock = 143,
-    ParachuteDeploy = 144,
-    ParachuteDetach = 145,
-    ParachuteTurnLR = 146,
-    ParachuteTurnLeftOnly = 147,
-    ParachuteTurnRightOnly = 148,
-    ParachutePitchUD = 149,
-    ParachutePitchUpOnly = 150,
-    ParachutePitchDownOnly = 151,
-    ParachuteBrakeLeft = 152,
-    ParachuteBrakeRight = 153,
-    ParachuteSmoke = 154,
-    ParachutePrecisionLanding = 155,
-    Map = 156,
-    SelectWeaponUnarmed = 157,
-    SelectWeaponMelee = 158,
-    SelectWeaponHandgun = 159,
-    SelectWeaponShotgun = 160,
-    SelectWeaponSMG = 161,
-    SelectWeaponAutoRifle = 162,
-    SelectWeaponSniper = 163,
-    SelectWeaponHeavy = 164,
-    SelectWeaponSpecial = 165,
-    SelectCharacterMichael = 166,
-    SelectCharacterFranklin = 167,
-    SelectCharacterTrevor = 168,
-    SelectCharacterMultiplayer = 169,
-    SaveReplayClip = 170,
-    SpecialAbilityPC = 171,
-    CellPhoneUp = 172,
-    CellPhoneDown = 173,
-    CellPhoneLeft = 174,
-    CellPhoneRight = 175,
-    CellPhoneSelect = 176,
-    CellPhoneCancel = 177,
-    CellPhoneOption = 178,
-    CellPhoneExtraOption = 179,
-    CellPhoneScrollForward = 180,
-    CellPhoneScrollBackward = 181,
-    CellPhoneCameraFocusLock = 182,
-    CellPhoneCameraGrid = 183,
-    CellPhoneCameraSelfie = 184,
-    CellPhoneCameraDOF = 185,
-    CellPhoneCameraExpression = 186,
-    FrontendDown = 187,
-    FrontendUp = 188,
-    FrontendLeft = 189,
-    FrontendRight = 190,
-    FrontendRDown = 191,
-    FrontendRUp = 192,
-    FrontendRLeft = 193,
-    FrontendRRight = 194,
-    FrontendAxisX = 195,
-    FrontendAxisY = 196,
-    FrontendRightAxisX = 197,
-    FrontendRightAxisY = 198,
-    FrontendPause = 199,
-    FrontendPauseAlternate = 200,
-    FrontendAccept = 201,
-    FrontendCancel = 202,
-    FrontendX = 203,
-    FrontendY = 204,
-    FrontendLB = 205,
-    FrontendRB = 206,
-    FrontendLT = 207,
-    FrontendRT = 208,
-    FrontendLS = 209,
-    FrontendRS = 210,
-    FrontendLeaderboard = 211,
-    FrontendSocialClub = 212,
-    FrontendSocialClubSecondary = 213,
-    FrontendDelete = 214,
-    FrontendEndscreenAccept = 215,
-    FrontendEndscreenExpand = 216,
-    FrontendSelect = 217,
-    ScriptLeftAxisX = 218,
-    ScriptLeftAxisY = 219,
-    ScriptRightAxisX = 220,
-    ScriptRightAxisY = 221,
-    ScriptRUp = 222,
-    ScriptRDown = 223,
-    ScriptRLeft = 224,
-    ScriptRRight = 225,
-    ScriptLB = 226,
-    ScriptRB = 227,
-    ScriptLT = 228,
-    ScriptRT = 229,
-    ScriptLS = 230,
-    ScriptRS = 231,
-    ScriptPadUp = 232,
-    ScriptPadDown = 233,
-    ScriptPadLeft = 234,
-    ScriptPadRight = 235,
-    ScriptSelect = 236,
-    CursorAccept = 237,
-    CursorCancel = 238,
-    CursorX = 239,
-    CursorY = 240,
-    CursorScrollUp = 241,
-    CursorScrollDown = 242,
-    EnterCheatCode = 243,
-    InteractionMenu = 244,
-    MPTextChatAll = 245,
-    MPTextChatTeam = 246,
-    MPTextChatFriends = 247,
-    MPTextChatCrew = 248,
-    PushToTalk = 249,
-    CreatorLs = 250,
-    CreatorRs = 251,
-    CreatorLt = 252,
-    CreatorRt = 253,
-    CreatorMenuToggle = 254,
-    CreatorAccept = 255,
-    CreatorDelete = 256,
-    Attack2 = 257,
-    RappelJump = 258,
-    RappelLongJump = 259,
-    RappelSmashWindow = 260,
-    PrevWeapon = 261,
-    NextWeapon = 262,
-    MeleeAttack1 = 263,
-    MeleeAttack2 = 264,
-    Whistle = 265,
-    MoveLeft = 266,
-    MoveRight = 267,
-    MoveUp = 268,
-    MoveDown = 269,
-    LookLeft = 270,
-    LookRight = 271,
-    LookUp = 272,
-    LookDown = 273,
-    SniperZoomIn = 274,
-    SniperZoomOut = 275,
-    SniperZoomInAlternate = 276,
-    SniperZoomOutAlternate = 277,
-    VehMoveLeft = 278,
-    VehMoveRight = 279,
-    VehMoveUp = 280,
-    VehMoveDown = 281,
-    VehGunLeft = 282,
-    VehGunRight = 283,
-    VehGunUp = 284,
-    VehGunDown = 285,
-    VehLookLeft = 286,
-    VehLookRight = 287,
-    ReplayStartStopRecording = 288,
-    ReplayStartStopRecordingSecondary = 289,
-    ScaledLookLr = 290,
-    ScaledLookUd = 291,
-    ScaledLookUpOnly = 292,
-    ScaledLookDownOnly = 293,
-    ScaledLookLeftOnly = 294,
-    ScaledLookRightOnly = 295,
-    ReplayMarkerDelete = 296,
-    ReplayClipDelete = 297,
-    ReplayPause = 298,
-    ReplayRewind = 299,
-    ReplayFfwd = 300,
-    ReplayNewmarker = 301,
-    ReplayRecord = 302,
-    ReplayScreenshot = 303,
-    ReplayHidehud = 304,
-    ReplayStartpoint = 305,
-    ReplayEndpoint = 306,
-    ReplayAdvance = 307,
-    ReplayBack = 308,
-    ReplayTools = 309,
-    ReplayRestart = 310,
-    ReplayShowhotkey = 311,
-    ReplayCyclemarkerleft = 312,
-    ReplayCyclemarkerright = 313,
-    ReplayFovincrease = 314,
-    ReplayFovdecrease = 315,
-    ReplayCameraup = 316,
-    ReplayCameradown = 317,
-    ReplaySave = 318,
-    ReplayToggletime = 319,
-    ReplayToggletips = 320,
-    ReplayPreview = 321,
-    ReplayToggleTimeline = 322,
-    ReplayTimelinePickupClip = 323,
-    ReplayTimelineDuplicateClip = 324,
-    ReplayTimelinePlaceClip = 325,
-    ReplayCtrl = 326,
-    ReplayTimelineSave = 327,
-    ReplayPreviewAudio = 328,
-    VehDriveLook = 329,
-    VehDriveLook2 = 330,
-    VehFlyAttack2 = 331,
-    RadioWheelUd = 332,
-    RadioWheelLr = 333,
-    VehicleSlowmoUd = 334,
-    VehSlowmoUpOnly = 335,
-    VehSlowmoDownOnly = 336,
-    VehHydraulicsControlToggle = 337,
-    VehHydraulicsControlLeft = 338,
-    VehHydraulicsControlRight = 339,
-    VehHydraulicsControlUp = 340,
-    VehHydraulicsControlDown = 341,
-    VehHydraulicsControlUD = 342,
-    VehHydraulicsControlLR = 343,
-    SwitchVisor = 344,
-    VehMeleeHold = 345,
-    VehMeleeLeft = 346,
-    VehMeleeRight = 347,
-    MapPOI = 348,
-    ReplaySnapmaticPhoto = 349,
-    VehCarJump = 350,
-    VehRocketBoost = 351,
-    VehFlyBoost = 352,
-    VehParachute = 353,
-    VehBikeWings = 354,
-    VehFlyBombBay = 355,
-    VehFlyCounter = 356,
-    VehTransform = 357
-}
-#endif
